milkit

+++ src/main/java/com/milkit_shop/controller/BoardController.java
@@ -3,7 +3,10 @@ package com.milkit_shop.controller;
 import com.milkit_shop.dto.BoardDto;
 import com.milkit_shop.dto.BoardReadDto;
 import com.milkit_shop.entity.Board;
+import com.milkit_shop.entity.Member;
 import com.milkit_shop.service.BoardService;
+import com.milkit_shop.service.MemberService;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,6 +19,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class BoardController {
   private final BoardService boardService;
+  private final MemberService memberService;
 
   @GetMapping("/mainboard")
   public ResponseEntity<List<BoardDto>> getBoard(@RequestParam(value = "search", required = false) String search){
@@ -38,13 +42,23 @@ public class BoardController {
   @PatchMapping("/{boardId}/delete")
   public ResponseEntity<Void> delete(@PathVariable Long boardId,Principal principal) {
     String email = principal.getName();
-    boardService.deleteBoard(boardId,email);
+    Member member=memberService.findMemberByEmail(email);
+    if(!boardService.duplicateBoard(boardId,email) && !member.getRole().name().equals("ROLE_ADMIN")) {
+      return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+    }else{
+      boardService.deleteBoard(boardId, email);
+    }
     return ResponseEntity.ok(null);
   }
   @PutMapping("/{boardId}/modify")
   public ResponseEntity<Void> update(@PathVariable Long boardId,@RequestBody BoardDto dto,Principal principal) {
     String email = principal.getName();
-    boardService.updateBoard(boardId,dto,email);
+    Member member=memberService.findMemberByEmail(email);
+    if(!boardService.duplicateBoard(boardId,email) && !member.getRole().name().equals("ROLE_ADMIN")) {
+      return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+    }else{
+      boardService.updateBoard(boardId,dto,email);
+    }
     return ResponseEntity.ok(null);
   }
